Backend/ExcelApp/requirements.txt
streamlit
pandas
openpyxl
mysql-connector-python
pdfplumber
xlrd
fuzzywuzzy
python-Levenshtein
